refactor(segmentation): replace `SRXUtils` with `SRXManager`
- Moved `SRXUtils` and related classes from `segmentation.util` to `segmentation`.
- Renamed `SRXUtils` to `SRXManager` for consistency with naming conventions.
- Updated all references and test cases to use `SRXManager`.
- Improved utility organization by consolidating segmentation logic into a single package.
