+++ src/org/omegat/util/Preferences.java
@@ -42,7 +42,7 @@ import java.util.Objects;
 
 import org.jetbrains.annotations.Nullable;
 import org.omegat.core.segmentation.SRX;
-import org.omegat.core.segmentation.util.SRXUtils;
+import org.omegat.core.segmentation.SRXManager;
 import org.omegat.filters2.master.FilterMaster;
 import org.omegat.filters2.master.FiltersUtil;
 
@@ -720,7 +720,7 @@ public final class Preferences {
 
         File srxDir = new File(StaticUtils.getConfigDir());
         try {
-            SRXUtils.saveToSrx(srx, srxDir); // save to segmentation.srx in the given
+            SRXManager.saveToSrx(srx, srxDir); // save to segmentation.srx in the given
                                         // directory
         } catch (IOException ex) {
             ex.printStackTrace();
@@ -827,7 +827,7 @@ public final class Preferences {
         didInitSegmentation = true;
 
         File srxDir = new File(StaticUtils.getConfigDir());
-        SRX s = SRXUtils.loadFromDir(srxDir); // may read SRX or CONF
+        SRX s = SRXManager.loadFromDir(srxDir); // may read SRX or CONF
         if (s == null) {
             s = SRX.getDefault();
         }
